Refactor ArrowKeyFocusManager
